# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShadowTrace",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/actions/{action_id}/approve": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Approve Action",
        "operationId": "approve_action_api_v1_actions__action_id__approve_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/{action_id}/reject": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Reject Action",
        "operationId": "reject_action_api_v1_actions__action_id__reject_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRejectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/{action_id}/resolve-unknown": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Resolve Unknown Action",
        "operationId": "resolve_unknown_action_api_v1_actions__action_id__resolve_unknown_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveUnknownRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{event_id}/dispositions": {
      "get": {
        "tags": [
          "dispositions"
        ],
        "summary": "List Event Dispositions",
        "operationId": "list_event_dispositions_api_v1_events__event_id__dispositions_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispositionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dispositions/{disposition_id}": {
      "get": {
        "tags": [
          "dispositions"
        ],
        "summary": "Get Disposition",
        "operationId": "get_disposition_api_v1_dispositions__disposition_id__get",
        "parameters": [
          {
            "name": "disposition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Disposition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/writebacks/{writeback_id}": {
      "get": {
        "tags": [
          "dispositions"
        ],
        "summary": "Get Writeback",
        "operationId": "get_writeback_api_v1_writebacks__writeback_id__get",
        "parameters": [
          {
            "name": "writeback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Writeback Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritebackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/writebacks/{writeback_id}/retry": {
      "post": {
        "tags": [
          "dispositions"
        ],
        "summary": "Retry Writeback",
        "operationId": "retry_writeback_api_v1_writebacks__writeback_id__retry_post",
        "parameters": [
          {
            "name": "writeback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Writeback Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritebackOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/writebacks/{writeback_id}/resolve": {
      "post": {
        "tags": [
          "dispositions"
        ],
        "summary": "Resolve Writeback",
        "operationId": "resolve_writeback_api_v1_writebacks__writeback_id__resolve_post",
        "parameters": [
          {
            "name": "writeback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Writeback Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveWritebackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritebackOperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionApproveRequest": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ActionApproveRequest"
      },
...
      "ActionOperationResponse": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "status"
        ],
        "title": "ActionOperationResponse"
      },
      "ActionRejectRequest": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ActionRejectRequest"
      },
...
      "ConfirmationEvidence": {
        "type": "string",
        "enum": [
          "adapter_acknowledged",
          "status_queried",
          "readback_verified",
          "manual_confirmed"
        ],
        "title": "ConfirmationEvidence"
      },
...
      "DispositionCommand": {
        "properties": {
          "disposition_id": {
            "type": "string",
            "title": "Disposition Id"
          },
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "closure_cycle": {
            "type": "integer",
            "title": "Closure Cycle"
          },
          "intent_kind": {
            "$ref": "#/components/schemas/DispositionIntentKind"
          },
          "source_locator": {
            "$ref": "#/components/schemas/SourceObjectLocator"
          },
          "operation_code": {
            "type": "string",
            "title": "Operation Code"
          },
          "operation_params": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SetEventDispositionParams"
              },
              {
                "$ref": "#/components/schemas/SubmitEntityActionParams"
              },
              {
                "$ref": "#/components/schemas/RecordExecutionResultParams"
              },
              {
                "$ref": "#/components/schemas/RecordCompensationParams"
              }
            ],
            "title": "Operation Params",
            "discriminator": {
              "propertyName": "operation_code",
              "mapping": {
                "record_compensation": "#/components/schemas/RecordCompensationParams",
                "record_execution_result": "#/components/schemas/RecordExecutionResultParams",
                "set_event_disposition": "#/components/schemas/SetEventDispositionParams",
                "submit_entity_action": "#/components/schemas/SubmitEntityActionParams"
              }
            }
          },
          "target_results": {
            "items": {
              "$ref": "#/components/schemas/TargetDispositionResult"
            },
            "type": "array",
            "title": "Target Results"
          },
          "operator_id": {
            "type": "string",
            "title": "Operator Id"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "source_concurrency_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Concurrency Token"
          },
          "execution_owner": {
            "$ref": "#/components/schemas/ExecutionOwner"
          },
          "parent_disposition_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Disposition Id"
          },
          "supersedes_disposition_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Disposition Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "disposition_id",
          "action_id",
          "closure_cycle",
          "intent_kind",
          "source_locator",
          "operation_code",
          "operation_params",
          "operator_id",
          "idempotency_key",
          "execution_owner"
        ],
        "title": "DispositionCommand",
        "description": "Minimal outbound disposition envelope (strict allowlist)."
      },
      "DispositionIntentKind": {
        "type": "string",
        "enum": [
          "entity_action_submit",
          "execution_result_record",
          "compensation_record",
          "event_status_update"
        ],
        "title": "DispositionIntentKind",
        "description": "Internal envelope classification; not a vendor enum."
      },
      "DispositionListResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DispositionResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "event_id"
        ],
        "title": "DispositionListResponse"
      },
...
      "DispositionResponse": {
        "properties": {
          "disposition": {
            "$ref": "#/components/schemas/DispositionCommand"
          },
          "writeback_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WritebackStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "disposition"
        ],
        "title": "DispositionResponse"
      },
...
      "ExecutionOwner": {
        "type": "string",
        "enum": [
          "xdr_managed",
          "direct_tool"
        ],
        "title": "ExecutionOwner"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecordCompensationParams": {
        "properties": {
          "operation_code": {
            "type": "string",
            "const": "record_compensation",
            "title": "Operation Code",
            "default": "record_compensation"
          },
          "summary_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RecordCompensationParams",
        "description": "Params for COMPENSATION_RECORD (rollback compensation sync)."
      },
      "RecordExecutionResultParams": {
        "properties": {
          "operation_code": {
            "type": "string",
            "const": "record_execution_result",
            "title": "Operation Code",
            "default": "record_execution_result"
          },
          "summary_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RecordExecutionResultParams",
        "description": "Params for EXECUTION_RESULT_RECORD (DIRECT_TOOL result sync)."
      },
...
      "ResolveUnknownRequest": {
        "properties": {
          "resolution": {
            "type": "string",
            "enum": [
              "mark_success",
              "mark_failed",
              "manual_confirmed"
            ],
            "title": "Resolution"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "evidence_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resolution",
          "comment"
        ],
        "title": "ResolveUnknownRequest"
      },
      "ResolveWritebackRequest": {
        "properties": {
          "resolution": {
            "type": "string",
            "enum": [
              "manual_confirmed",
              "mark_failed",
              "abandon"
            ],
            "title": "Resolution"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "evidence_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resolution",
          "comment"
        ],
        "title": "ResolveWritebackRequest"
      },
...
      "SetEventDispositionParams": {
        "properties": {
          "operation_code": {
            "type": "string",
            "const": "set_event_disposition",
            "title": "Operation Code",
            "default": "set_event_disposition"
          },
          "target_disposition": {
            "$ref": "#/components/schemas/SourceDisposition"
          },
          "comment_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_disposition"
        ],
        "title": "SetEventDispositionParams",
        "description": "Params for EVENT_STATUS_UPDATE (the deferred terminal disposition)."
      },
...
      "SourceObjectLocator": {
        "properties": {
          "source_product": {
            "type": "string",
            "title": "Source Product"
          },
          "source_tenant_id": {
            "type": "string",
            "title": "Source Tenant Id"
          },
          "connector_id": {
            "type": "string",
            "title": "Connector Id"
          },
          "source_kind": {
            "$ref": "#/components/schemas/SourceObjectKind"
          },
          "source_object_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Object Type"
          },
          "source_object_id": {
            "type": "string",
            "title": "Source Object Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_product",
          "source_tenant_id",
          "connector_id",
          "source_kind",
          "source_object_id"
        ],
        "title": "SourceObjectLocator",
        "description": "Minimal writeback locator (intro §4.3.5 / ISSUE-002)."
      },
...
      "SubmitEntityActionParams": {
        "properties": {
          "operation_code": {
            "type": "string",
            "const": "submit_entity_action",
            "title": "Operation Code",
            "default": "submit_entity_action"
          },
          "entity_action_code": {
            "type": "string",
            "title": "Entity Action Code"
          },
          "canonical_target": {
            "type": "string",
            "title": "Canonical Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entity_action_code",
          "canonical_target"
        ],
        "title": "SubmitEntityActionParams",
        "description": "Params for ENTITY_ACTION_SUBMIT (XDR_MANAGED entity action)."
      },
...
      "TargetWritebackResult": {
        "properties": {
          "canonical_target": {
            "type": "string",
            "title": "Canonical Target"
          },
          "status": {
            "$ref": "#/components/schemas/TargetWritebackStatus"
          },
          "provider_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Code"
          },
          "message_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Code"
          },
          "artifact_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "canonical_target",
          "status"
        ],
        "title": "TargetWritebackResult",
        "description": "Per-target writeback result (kept separate from outbound disposition result)."
      },
      "TargetWritebackStatus": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "confirmed",
          "failed",
          "conflict",
          "unknown"
        ],
        "title": "TargetWritebackStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WritebackOperationResponse": {
        "properties": {
          "writeback_id": {
            "type": "string",
            "title": "Writeback Id"
          },
          "status": {
            "$ref": "#/components/schemas/WritebackStatus"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "writeback_id",
          "status"
        ],
        "title": "WritebackOperationResponse"
      },
...
      "WritebackResponse": {
        "properties": {
          "writeback_id": {
            "type": "string",
            "title": "Writeback Id"
          },
          "disposition_id": {
            "type": "string",
            "title": "Disposition Id"
          },
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "status": {
            "$ref": "#/components/schemas/WritebackStatus"
          },
          "confirmation_evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmationEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "strong",
                  "medium",
                  "weak"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Tier"
          },
          "provider_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Code"
          },
          "message_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Code"
          },
          "target_results": {
            "items": {
              "$ref": "#/components/schemas/TargetWritebackResult"
            },
            "type": "array",
            "title": "Target Results"
          }
        },
        "type": "object",
        "required": [
          "writeback_id",
          "disposition_id",
          "action_id",
          "status"
        ],
        "title": "WritebackResponse",
        "description": "Redacted writeback view. ``raw_result`` is intentionally not exposed."
      },
      "WritebackStatus": {
        "type": "string",
        "enum": [
          "pending",
          "sending",
          "accepted",
          "confirmed",
          "partial",
          "failed",
          "conflict",
          "unknown"
        ],
        "title": "WritebackStatus",
        "description": "Only valid once a writeback command exists; null otherwise."
      }
    }
  }
}